requirements.txt
torch==2.2.2
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
networkx==3.3
PyYAML==6.0.1
scipy==1.13.1
tqdm==4.66.4
typer==0.12.3
